fix(paymentmodal clientstable): fix the amunts and add discount to the final vrew

+++ src/components/CurrencySpan.tsx
@@ -12,7 +12,7 @@ const CurrencySpan = ({
       {new Intl.NumberFormat('es-MX', {
         style: 'currency',
         currency: 'MXN'
-      }).format(parseInt(`${quantity}`))}{' '}
+      }).format(parseFloat(`${quantity}`))}{' '}
     </span>
   )
 }
